requirements.txt
autopep8==1.3.5
bcrypt==3.1.4
blinker==1.4
click==6.7
flake8==3.5.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
flask-shell-ipython==0.3.0
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.2.1
Faker==0.9.0
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.13
Pillow==5.1.0
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
pylint==1.8.4
python-dotenv==0.8.2
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
